+++ dmr/_compiled/negotiation.py
@@ -22,13 +22,12 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-
 import re
 from collections.abc import Iterable
 from typing import Final, final
 
 
-def accepted_type(
+def accepted_type(  # noqa: C901
     accept_value: str,
     provided_types: Iterable[str],
 ) -> str | None:
@@ -48,40 +47,109 @@ def accepted_type(
         Otherwise the provided type is returned as-is.
 
     """
-    accepted_types = sorted(
-        [_MediaTypeHeader(typ) for typ in accept_value.split(',')],
-        key=_by_priority,
-        reverse=True,
-    )
+    if not accept_value:
+        return None
 
-    types = [_MediaTypeHeader(typ) for typ in provided_types]
+    types = [_MediaTypeHeader(typ) for typ in provided_types if typ]
 
-    for accepted in accepted_types:
+    if not types:
+        return None
+
+    if ',' in accept_value:
+        accepted_types = [
+            _MediaTypeHeader(typ) for typ in accept_value.split(',') if typ
+        ]
+        accepted_types.sort(
+            key=lambda media: media.priority,
+            reverse=True,
+        )
+        for accepted in accepted_types:
+            for provided in types:
+                if provided.match(accepted):
+                    # Return the accepted type with wildcards replaced
+                    # by concrete parts from the provided type:
+                    return provided.as_string(
+                        accepted.maintype,
+                        accepted.subtype,
+                    )
+    else:
+        accepted = _MediaTypeHeader(accept_value)
         for provided in types:
-            if provided.match(accepted):
+            if provided.match(accepted):  # noqa: WPS441
                 # Return the accepted type with wildcards replaced
                 # by concrete parts from the provided type:
-                return provided.as_string(accepted.maintype, accepted.subtype)
+                return provided.as_string(accepted.maintype, accepted.subtype)  # noqa: WPS441
+
     return None
+
+
+def accepted_header(accept_value: str, media_type: str) -> bool:  # noqa: C901
+    """
+    Does the client accept a response in the given media type?
+
+    This is a faster alternative to Django's ``HttpRequest.accepts``.
+
+    Args:
+        accept_value: The value of ``Accept`` header.
+        media_type: The media type to check, e.g. ``"application/json"``.
+
+    Returns:
+        ``True`` if the media type is accepted according to the
+        ``Accept`` header, otherwise ``False``.
+
+    For example:
+
+        .. code:: python
+
+            >>> from django.http import HttpRequest
+            >>> request = HttpRequest()
+            >>> request.META = {'HTTP_ACCEPT': 'application/json'}
+            >>> accepted_header(
+            ...     request.headers.get('Accept'), 'text/plain'
+            ... )  # equivalent to request.accepts("text/plain")
+            False
+            >>> accepted_header(
+            ...     'application/json,text/html;q=0.8',
+            ...     'application/json',
+            ... )  # can be called with any headers-like mapping
+            True
+
+    """
+    if not accept_value or not media_type:
+        return False
+
+    provided = _MediaTypeHeader(media_type)
+
+    if ',' in accept_value:
+        for typ in accept_value.split(','):
+            if not typ:
+                continue
+            if provided.match(_MediaTypeHeader(typ)):
+                return True
+    elif provided.match(_MediaTypeHeader(accept_value)):
+        return True
+
+    return False
 
 
 @final
 class _MediaTypeHeader:
     """A helper class for ``Accept`` header parsing."""
 
-    __slots__ = ('maintype', 'params_str', 'priority', 'qparams', 'subtype')
+    __slots__ = ('maintype', 'params_str', 'qparams', 'subtype')
 
     def __init__(self, type_str: str) -> None:
         # preserve the original parameters, because the order might be
         # changed in the dict
-        self.params_str = ''.join(type_str.partition(';')[1:])
+        self.params_str = (
+            f';{type_str.partition(";")[2]}' if ';' in type_str else ''  # noqa: WPS237
+        )
 
         full_type, qparams = _parse_content_header(type_str)
         self.qparams = qparams
         maintype, _, subtype = full_type.partition('/')
         self.maintype = maintype
         self.subtype = subtype
-        self.priority = self._get_priority()
 
     def match(self, other: '_MediaTypeHeader') -> bool:
         for key, param_value in self.qparams.items():
@@ -105,7 +173,8 @@ class _MediaTypeHeader:
         subtype = subtype if self.subtype == '*' else self.subtype
         return f'{maintype}/{subtype}{self.params_str}'
 
-    def _get_priority(self) -> tuple[int, int]:  # noqa: WPS231
+    @property  # don't use cached_property since it's accessed only once
+    def priority(self) -> tuple[int, int]:
         # Use fixed point values with two decimals to avoid problems
         # when comparing float values
         quality = 100
@@ -129,10 +198,6 @@ class _MediaTypeHeader:
             specificity = 3
 
         return quality, specificity
-
-
-def _by_priority(media: _MediaTypeHeader) -> tuple[int, int]:
-    return media.priority
 
 
 _token: Final = r"([\w!#$%&'*+\-.^_`|~]+)"  # noqa: S105